feat: Arena auto generation
I haven't tested bc im too lazy to setup mongo so imma trust the existing location calculations in neptune.

This should ensure that the clipboard for the arena will always be pasted from same point so you should be able to just add the offsets and get a perfectly working arena.